mindahin ke vc baru dan add new swift file
